refactor: change organization of pages to enable possibility of layout differences
