+++ apps/bytebank/src/app/(pages)/(financial-pages)/_components/page.module.scss
@@ -2,16 +2,10 @@
     display: flex;
     flex-direction: column;
     gap: 16px;
-    width: 100%;          
-    max-width: 690px;     // Mas nunca ultrapassa 690px
-    background-color: white;
-    border-radius: 8px;
-    padding-bottom: 15px;
-    margin-bottom: 24px;
+    width: 100%;    
     
 /* Para desktop, quando a largura é maior que 1024px */
 @media (min-width: 1025px) {
-    width: 282px; 
     top: 0;
     right: 0;
 }
