+++ apps/bytebank/src/app/(pages)/(financial-pages)/page.module.scss
@@ -4,7 +4,6 @@
   align-items: center;
   min-height: 100vh;       
   width: 100%;
-
   box-sizing: border-box;
 }
 
@@ -15,8 +14,6 @@
     justify-content: center; 
     align-items: flex-start;
     gap: 24px;
-    margin-top: 24px;
-    margin-bottom: 24px;
     width: 100%;
     box-sizing: border-box;
 
@@ -35,8 +32,6 @@
     display: flex;
     flex-direction: column;
     gap: 24px;
-    flex: 1;
-    max-width: 600px;
 
     @media (max-width: 1023px) {
         width: 100%;
@@ -45,9 +40,6 @@
 }
 
 .rightColumn {
-    flex: 1;
-    max-width: 500px;
-
     @media (max-width: 1023px) {
         width: 100%;
         max-width: 100%;
