+++ apps/bytebank/src/app/(pages)/_components/card-transaction/style.scss
@@ -3,7 +3,6 @@
   background-image: 
   url('/images/pixels-04.png'), 
   url('/images/pixels-03.png') !important;
-  max-width: 690px;
   background-position: 
   right top, 
   left bottom !important;
